add forgot password, fix translation, create email

+++ src/Controller/AccountController.php
@@ -12,6 +12,11 @@ use App\Form\NewPasswordFormType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use App\Entity\User;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
+
+use App\Security\EmailVerifier;
+use Symfony\Bridge\Twig\Mime\TemplatedEmail;
+use Symfony\Component\Mime\Address;
 
 
 /**
@@ -20,10 +25,16 @@ use App\Entity\User;
 class AccountController extends AbstractController
 {
     /**
-     * @Route("/", name="app_account", methods="GET")
+     * @Route("/", name="app_account_index", methods="GET")
      */
-    public function show()
+    public function show(SessionInterface $session)
     {
+        if (!$this->getUser()) 
+            return $this->redirectToRoute('app_info_home');    
+        elseif(!$this->getUser()->getIsVerified() && !$session->get('emailSent'))
+            $this->addFlash('error', '<span class="h5">Votre compte n\'est pas verifié.<span> <a href=".\email" class="h5 text-light font-weight-bold">Recevoir un email pour le verifier</a>');
+        
+
         return $this->render('account/show.html.twig', [      
         ]);
     }
@@ -33,6 +44,7 @@ class AccountController extends AbstractController
      */
     public function edit(Request $request, EntityManagerInterface $em, UserPasswordEncoderInterface $passwordEncoder) : Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         if($this->isCsrfTokenValid('account_edit', $request->query->get('token'))) {
             $user = $this->getUser();
             $orginalMail = $user->getEmail();
@@ -54,7 +66,7 @@ class AccountController extends AbstractController
 
                 $this->addFlash('success', 'Votre compte a été mis à jour avec succès');
 
-                return $this->redirectToRoute('app_account');
+                return $this->redirectToRoute('app_account_index');
             }
         }
         else
@@ -71,6 +83,8 @@ class AccountController extends AbstractController
      */
     public function NewPassword(Request $request, EntityManagerInterface $em, UserPasswordEncoderInterface $passwordEncoder) : Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         if($this->isCsrfTokenValid('account_NewPassword', $request->query->get('token'))) {
             $user = $this->getUser();
 
@@ -91,7 +105,7 @@ class AccountController extends AbstractController
 
                 $this->addFlash('success', 'Votre mot de passe a été mis à jour avec succès');
 
-                return $this->redirectToRoute('app_account');
+                return $this->redirectToRoute('app_account_index');
             }
         }
         else
@@ -103,10 +117,40 @@ class AccountController extends AbstractController
     }
 
     /**
+     * @Route("/email", name="app_account_email", methods={"GET"})
+     */
+    public function emailValidation(EmailVerifier $emailVerifier, SessionInterface $session): Response
+    {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+        if(!$this->getUser()->getIsVerified() && !$session->get('emailSent'))
+        {
+             $emailVerifier->sendEmailConfirmation('app_verify_email', $this->getUser(),
+            (new TemplatedEmail())
+                ->from(new Address($this->getParameter('app.mail_from_address'), $this->getParameter('app.mail_from_name')))
+                ->to($this->getUser()->getEmail())
+                ->subject('Please Confirm your Email')
+                ->htmlTemplate('mails/confirmation.html.twig')
+            );
+
+            $session->set('emailSent', true);
+            $this->addFlash('success', 'Un email vous a été envoyé !');
+        }
+        else{
+            $this->addFlash('info', 'Un email vous déjà a été envoyé.');
+        }   
+       
+        return $this->redirectToRoute('app_account_index');
+    }
+
+
+    /**
      * @Route("/delete", name="app_account_delete", methods={"DELETE"})
      */
     public function delete(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         if ($this->isCsrfTokenValid('delete', $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($this->getUser());
